Create lxqt-spike-meta.ebuild
lxqt needs a slight sed ++cron added as most users wont run the menu updater. ie on load refresh menus
similar to xfce or lxde in layout like enlightenment it's extremely fast. or add a desktop Link to run shell to do menus for time being.
for now seeing if it builds with the spike toys.